webcolegio/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
colorama==0.4.1
Django==2.2.3
django-ckeditor==5.7.1
django-js-asset==1.2.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
